+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Introducing `v1alpha3` CR's.
+
 ### Added
 
 - Add check to ensure that the `Cluster` resource is in the same namespace as the `certConfig` before creating the secret there.