+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.0] - 2021-10-15
+
 ### Changed
 
 - Introducing `v1alpha3` CR's.
@@ -77,7 +79,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - First release.
 
-[Unreleased]: https://github.com/giantswarm/cert-operator/compare/v1.1.0...HEAD
+[Unreleased]: https://github.com/giantswarm/cert-operator/compare/v1.2.0...HEAD
+[1.2.0]: https://github.com/giantswarm/cert-operator/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/giantswarm/cert-operator/compare/v1.0.1...v1.1.0
 [1.0.1]: https://github.com/giantswarm/cert-operator/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/giantswarm/cert-operator/compare/v0.1.0-2...v1.0.0